requirements.txt
anthropic>=0.88.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
supabase>=2.0.0
twilio>=9.0.0
stripe>=7.0.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
weasyprint>=60.0
requests>=2.31.0
aiohttp>=3.9.0
python-multipart>=0.0.6
jinja2>=3.1.0
pgvector>=0.2.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0
httpx>=0.25.0
